crypto-dataset/requirements.txt
pycryptodome
numpy
pandas
scikit-learn
joblib
tqdm
fastapi
uvicorn
python-multipart
